create a post api route

+++ api/controllers/user.controller.js
@@ -10,6 +10,9 @@ export const test = (req, res) => {
 
 //update user
 export const updateUser = async (req, res, next) => {
+  if (req.user.id !== req.params.userId) {
+    return next(errorHandler(403, "Access Denied"));
+  }
   if (req.body.password) {
     if (req.body.password.length < 6) {
       return next(
@@ -68,6 +71,9 @@ export const updateUser = async (req, res, next) => {
 // delete user
 export const deleteUser = async (req, res, next) => {
   try{
+    if (req.user.id !== req.params.userId) {
+      return next(errorHandler(403, "Access Denied"));
+    }
     await User.findByIdAndDelete(req.params.userId);
     return res.status(200).json({message: "User deleted successfully"});
   }
